// VIS_Application/app/Model/Notification/MyTicket.ts
export interface IMyTicket {
    CompanyId: number;
    Mode: string;
    Id: number;
    Subject: string;  
    CreatedOn: Date;
    AddressToGroup: string;
    AssignTo: string;
    CreatedByName: string;
    UpdatedByName: string;
    UpdatedOn: Date;
    Priority: string;
    Status: string;
    CreatedBy: number; 
    UpdatedBy: number
    IsActive: boolean;
    CreatedId: number;
    FileName: string;
    HelpTicketId: number;
    OriginalFileName: string;

}